# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontosyx API",
    "description": "The Semantic Orchestrator — Knowledge Graph Lifecycle Platform.\n\n## Response envelope\n\nEvery successful 2xx response is wrapped in:\n\n```json\n{ \"data\": <T>, \"pagination\": {\"next_cursor\": \"...\"}?, \"meta\": {...}? }\n```\n\nEach handler's `responses(... body = T)` documents the type of `data`, **not** the wire shape. Generated clients should either unwrap `data` in a post-processing step (the approach `web/src/lib/api/client.ts` takes — frontend callers receive `T` directly) or wrap every response type in the envelope shape when consuming the spec from a third-party codegen.\n\nError responses use a separate envelope: `{ \"error\": { \"type\": \"...\", \"message\": \"...\" } }` (see the `ErrorResponse` schema).",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects/{id}/revisions": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "list_revisions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of ontology revision snapshots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/projects/{id}/revisions/{rev}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "get_revision",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "rev",
            "in": "path",
            "description": "Revision number",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ontology revision snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Revision not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/projects/{id}/revisions/{rev}/restore": {
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "restore_revision",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "rev",
            "in": "path",
            "description": "Revision number to restore",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revision restored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRestoreResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project or revision not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorBody"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "details": {
            "description": "Optional additional details"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "type": {
            "type": "string",
            "description": "Machine-readable error type (e.g., \"not_found\", \"bad_request\")"
          }
        }
      },
...
      "ProjectRestoreResponse": {
        "type": "object",
        "required": [
          "project"
        ],
        "properties": {
          "project": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
...
}